pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nodiview"
version = "0.1.0"
description = "Ein moderner Linux-Bildbetrachter mit integrierter Bildoptimierung"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0"}
authors = [
    {name = "nodiView Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Viewers",
]
dependencies = [
    "PyGObject>=3.42.0",
    "Pillow>=10.0.0",
    "pyvips>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
nodiview = "nodiview.main:main"

[tool.setuptools]
packages = ["nodiview", "nodiview.optimizer", "nodiview.converter", "nodiview.editor", "nodiview.batch", "nodiview.utils"]

[tool.setuptools.package-data]
nodiview = ["data/icons/*", "data/*.desktop", "data/*.appdata.xml"]

[tool.black]
line-length = 100
target-version = ['py310']

